refactor: convert a bunch of files to typescript
