+++ packages/saber/.eslintignore
@@ -1,2 +1,3 @@
 **/vendor/**
 *.min.js
+**/types/**/*.d.ts
