+++ packages/saber/.eslintrc.js
@@ -1,6 +1,7 @@
 module.exports = {
   extends: 'saber',
   rules: {
-    '@typescript-eslint/triple-slash-reference': 'off'
+    '@typescript-eslint/triple-slash-reference': 'off',
+    '@typescript-eslint/no-explicit-any': 'off'
   }
 }
