--- packages/saber/declarations.d.ts
@@ -1,3 +1,0 @@
-declare type TODO = any
-
-declare type NonUndefined<A> = A extends undefined ? never : A
