+++ packages/saber/example/pages/wow.js
@@ -6,7 +6,7 @@ export default {
   }
 }
 
-export const attributes = {
+export const data = {
   title: 'wow'
 }
