+++ packages/saber/src/utils/getFileNames.ts
@@ -1,8 +1,4 @@
-// @ts-check
-/**
- * @param {string} useHash
- */
-module.exports = useHash => {
+export default (useHash: boolean) => {
   return {
     js: useHash ? 'js/[name].[chunkhash:8].js' : 'js/[name].js',
     css: useHash ? 'css/[name].[chunkhash:8].css' : 'css/[name].css',
