+++ packages/saber/src/utils/getPageType.ts
@@ -1,0 +1,7 @@
+export default (relative: string) => {
+  if (relative.startsWith('_posts/')) {
+    return 'post'
+  }
+
+  return 'page'
+}
