+++ packages/saber/src/utils/getPermalink.ts
@@ -1,10 +1,10 @@
-// @ts-check
+import { Page } from '../Pages'
+import { Permalinks } from '../types'
 
 /**
  * Left-pad '0' to number that is smaller than 10
- * @param {number} input
  */
-function padZero(input) {
+function padZero(input: number) {
   return input < 10 ? `0${input}` : input
 }
 
@@ -15,14 +15,12 @@ function padZero(input) {
 
 /**
  * Get the permalink of a page
- * @param {string[]} localeNames
- * @param {object} page - Page object
- * @param {string} page.slug - The page slug
- * @param {string} page.type - The page type
- * @param {Date=}  page.createdAt - The creation time
- * @param {any} permalinks - The `permalinks` config option
  */
-module.exports = (localeNames, page, permalinks) => {
+export default (
+  localeNames: string[],
+  page: { slug?: string; createdAt?: Date; type: 'page' | 'post' },
+  permalinks: Permalinks
+) => {
   if (page.slug === 'index') return '/'
 
   permalinks = Object.assign(
@@ -37,7 +35,7 @@ module.exports = (localeNames, page, permalinks) => {
   const date = page.createdAt && new Date(page.createdAt)
 
   let prefix = ''
-  let { slug } = page
+  let { slug = '' } = page
   for (const localeName of localeNames) {
     const RE = new RegExp(`^${localeName}(/|$)`)
     if (RE.test(slug)) {
